Split `sign.rs` into modules
